feat: add cffi feature
When building the wasm from `flux-lsp`, the C extern functions are still
provided as part of the public interface for the wasm, despite the
library not ever using them. By putting this behind a (default) feature
flag, and then depending and this crate without the default features, we
trim about 200k from the final wasm binary size.
