+++ libflux/go/libflux/buildinfo.gen.go
@@ -58,10 +58,11 @@ var sourceHashes = map[string]string{
 	"libflux/flux-core/src/semantic/walk/mod.rs":                                                  "8f6bcfc7d309ad821016d6622b49cfde3de5de4d7d0315e55c5cd7cfed279eb8",
 	"libflux/flux-core/src/semantic/walk/test_utils.rs":                                           "b980587707038c420d16b99f99587e69b71f02c32277db37c7e4a094d32f2b38",
 	"libflux/flux-core/src/semantic/walk/walk_mut.rs":                                             "4ef980dd4e64a4522d1d79d7bb72f5e3743c3e092387d46c8295d9364b5e2c12",
-	"libflux/flux/Cargo.toml":                                                                     "6152dbad3f3e73120a54088b207d37ad1bd61cdf7abca6be27364dc7b809594b",
+	"libflux/flux/Cargo.toml":                                                                     "f9f73d9176225e75b2bf658dce4b1e4294db8a451b96ba28323afad4840cd54b",
 	"libflux/flux/FLUXDOC.md":                                                                     "92e6dd8043bd87b4924e09aa28fb5346630aee1214de28ea2c8fc0687cad0785",
 	"libflux/flux/build.rs":                                                                       "eea9cfad417042bdd0908c5f7bcf2f004ee70031d973292e2cf54fab526a1807",
-	"libflux/flux/src/lib.rs":                                                                     "de4e758c4e5a639e5a59d591074ef4580e2a0d203fd28469bde32a3582bf32be",
+	"libflux/flux/src/cffi.rs":                                                                    "b3d57670725039f848c414e0f796c435599c690bef7e6c8f7f4d5a3cc1373401",
+	"libflux/flux/src/lib.rs":                                                                     "32d305d27a04cd061b796c0f70bedbae9cb4fcdbcc7d08505c74cf9c22adc953",
 	"libflux/flux/templates/base.html":                                                            "a818747b9621828bb96b94291c60922db54052bbe35d5e354f8e589d2a4ebd02",
 	"libflux/flux/templates/home.html":                                                            "f9927514dd42ca7271b4817ad1ca33ec79c03a77a783581b4dcafabd246ebf3f",
 	"libflux/flux/templates/package.html":                                                         "635e1d638ad1a430f3894ff0a458572ca3c4dc6b9cec2c57fe771443849e1402",
